HelperFunctions/requirements.txt
dash==2.2.0
imagesize==1.3.0
joblib==1.1.0
numpy==1.19.5
opencv_contrib_python==4.5.5.62
pandas==1.1.5
Pillow==9.1.1
plotly==5.6.0
pyperclip==1.8.2
scikit_learn==1.1.1
seaborn==0.11.2
